+++ app/javascript/frontend/pages/Fleets/Show/index.vue
@@ -174,9 +174,9 @@
       <template slot="actions">
         <BtnDropdown size="small">
           <template v-if="mobile">
-            <Btn :to="{ name: 'fleet-stats' }" size="small" variant="link">
+            <Btn :to="{ name: 'fleet-stats' }" size="small" variant="dropdown">
               <i class="fad fa-chart-bar" />
-              {{ $t('labels.fleetStats') }}
+              <span>{{ $t('labels.fleetStats') }}</span>
             </Btn>
 
             <hr />
@@ -186,21 +186,21 @@
             :active="detailsVisible"
             :aria-label="toggleDetailsTooltip"
             size="small"
-            variant="link"
+            variant="dropdown"
             @click.native="toggleDetails"
           >
             <i class="fad fa-info-square" />
-            {{ toggleDetailsTooltip }}
+            <span>{{ toggleDetailsTooltip }}</span>
           </Btn>
 
-          <Btn size="small" variant="link" @click.native="toggleGrouped">
+          <Btn size="small" variant="dropdown" @click.native="toggleGrouped">
             <template v-if="grouped">
               <i class="fas fa-square" />
-              {{ $t('actions.ungrouped') }}
+              <span>{{ $t('actions.ungrouped') }}</span>
             </template>
             <template v-else>
               <i class="fas fa-th-large" />
-              {{ $t('actions.groupedByModel') }}
+              <span>{{ $t('actions.groupedByModel') }}</span>
             </template>
           </Btn>
         </BtnDropdown>