+++ app/javascript/frontend/pages/Hangar/Fleetchart/index.vue
@@ -48,7 +48,7 @@
             <Starship42Btn
               :vehicles="records"
               size="small"
-              variant="link"
+              variant="dropdown"
               :with-icon="true"
             />
 
@@ -59,12 +59,12 @@
             element="#fleetchart"
             filename="my-hangar-fleetchart"
             size="small"
-            variant="link"
+            variant="dropdown"
           />
 
           <hr />
 
-          <FleetChartStatusBtn size="small" variant="link" />
+          <FleetChartStatusBtn size="small" variant="dropdown" />
         </BtnDropdown>
       </template>